+++ acropolis/cascade.py
@@ -1,7 +1,5 @@
 # math
 from math import pi, log, log10, exp, sqrt
-# functools
-from functools import wraps
 # numpy
 import numpy as np
 # scipy
@@ -15,6 +13,8 @@ import warnings
 # db
 from .db import import_data_from_db
 from .db import in_rate_db, interp_rate_db
+# cache
+from .cache import cached_member
 # pprint
 from .pprint import print_warning, print_error
 # params
@@ -23,31 +23,6 @@ from .params import zeta3, pi2
 from .params import Emin
 from .params import approx_zero, eps, Ephb_T_max, E_EC_cut
 from .params import NE_pd, NE_min
-
-
-def _cached(f_uncached):
-    # Define the cache as a dictionary
-    cache = {}
-    cT = {"_": -1.}
-
-    # Define the wrapper function
-    @wraps(f_uncached)
-    def f_cached(*args):
-        T     = args[-1]
-        pargs = args[1:]
-
-        # For each new temperature,
-        # clear the cache and start over
-        if T != cT["_"]:
-            cT["_"] = T
-            cache.clear()
-
-        if pargs not in cache:
-            cache[pargs] = f_uncached(*args)
-
-        return cache[pargs]
-
-    return f_cached
 
 
 # _ReactionWrapperScaffold ####################################################
@@ -385,7 +360,7 @@ class _PhotonReactionWrapper(_ReactionWrapperScaffold):
 
 
     # INVERSE COMPTON SCATTERING ##############################################
-    @_cached
+    @cached_member
     def _kernel_inverse_compton(self, E, Ep, T):
         # Incorporate the non-generic integration limit as
         # the algorithm requires Ep > E and not Ep > E + me
@@ -445,7 +420,7 @@ class _ElectronReactionWrapper(_ReactionWrapperScaffold):
     # T is the temperature of the background photons
 
     # INVERSE COMPTON SCATTERING ##############################################
-    @_cached
+    @cached_member
     def _rate_inverse_compton(self, E, T):
         # Define the upper limit for the integration over x
         ulim = min( E - me2/(4.*E), Ephb_T_max*T )
@@ -483,7 +458,7 @@ class _ElectronReactionWrapper(_ReactionWrapperScaffold):
     # T  is the temperature of the background photons
 
     # INVERSE COMPTON SCATTERING ##############################################
-    @_cached
+    @cached_member
     def _kernel_inverse_compton(self, E, Ep, T):
         # E == Ep leads to a divergence in
         # the Bose-Einstein distribution
@@ -541,7 +516,7 @@ class _ElectronReactionWrapper(_ReactionWrapperScaffold):
 
 
     # BETHE_HEITLER PAIR CREATION #############################################
-    @_cached
+    @cached_member
     def _kernel_bethe_heitler(self, E, Ep, T):
         # Incorporate the non-generic integration limit as
         # the algorithm requires Ep > E and not Ep > E + me
@@ -553,7 +528,7 @@ class _ElectronReactionWrapper(_ReactionWrapperScaffold):
 
 
     # DOUBLE PHOTON PAIR CREATION #############################################
-    @_cached
+    @cached_member
     def _kernel_pair_creation(self, E, Ep, T):
         # In general, the threshold is Ep >~ me^2/(22*T)
         # However, here we use a slighlty smaller threshold
